модуль category_iterator и test_category_iterator
