+++ htmlcov/class_index.html
@@ -55,7 +55,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
 </header>
@@ -129,32 +129,40 @@
                 <td class="right" data-ratio="13 13">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t12">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t12"><data value='CategoryIterator'>CategoryIterator</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t6">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t6"><data value='CategoryIterator'>CategoryIterator</data></a></td>
                 <td>10</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="10 10">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t39">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t39"><data value='Smartphone'>Smartphone</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
+                <td>6</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="6 6">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t10">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t10"><data value='Smartphone'>Smartphone</data></a></td>
                 <td>5</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="5 5">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t67">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t67"><data value='LawnGrass'>LawnGrass</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t38">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t38"><data value='LawnGrass'>LawnGrass</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t91">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t91"><data value='Order'>Order</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t62">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t62"><data value='Order'>Order</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
@@ -163,10 +171,10 @@
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html"><data value=''><span class='no-noun'>(no class)</span></data></a></td>
-                <td>55</td>
+                <td>50</td>
                 <td>29</td>
                 <td>0</td>
-                <td class="right" data-ratio="26 55">47%</td>
+                <td class="right" data-ratio="21 50">42%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html#t4">src\print_mixin.py</a></td>
@@ -213,10 +221,10 @@
             <tr class="total">
                 <td class="name left">Total</td>
                 <td class="name left">&nbsp;</td>
-                <td>195</td>
+                <td>196</td>
                 <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="154 195">79%</td>
+                <td class="right" data-ratio="155 196">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -228,7 +236,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
     <aside class="hidden">
