+++ htmlcov/function_index.html
@@ -55,7 +55,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
 </header>
@@ -169,64 +169,72 @@
                 <td class="right" data-ratio="13 13">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t15">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t15"><data value='init__'>CategoryIterator.__init__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t9">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t9"><data value='init__'>CategoryIterator.__init__</data></a></td>
                 <td>3</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="3 3">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t22">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t22"><data value='iter__'>CategoryIterator.__iter__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t16">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t16"><data value='iter__'>CategoryIterator.__iter__</data></a></td>
                 <td>2</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="2 2">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t28">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t28"><data value='next__'>CategoryIterator.__next__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t22">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html#t22"><data value='next__'>CategoryIterator.__next__</data></a></td>
                 <td>5</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="5 5">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t47">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t47"><data value='init__'>Smartphone.__init__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html">src\category_iterator.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html"><data value=''><span class='no-noun'>(no function)</span></data></a></td>
+                <td>6</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="6 6">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t18">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t18"><data value='init__'>Smartphone.__init__</data></a></td>
                 <td>5</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="5 5">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t74">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t74"><data value='init__'>LawnGrass.__init__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t45">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t45"><data value='init__'>LawnGrass.__init__</data></a></td>
                 <td>4</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t94">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t94"><data value='init__'>Order.__init__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t65">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t65"><data value='init__'>Order.__init__</data></a></td>
                 <td>2</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="2 2">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t101">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t101"><data value='total_price'>Order.total_price</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t72">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t72"><data value='total_price'>Order.total_price</data></a></td>
                 <td>1</td>
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="1 1">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t106">src\main.py</a></td>
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t106"><data value='str__'>Order.__str__</data></a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t77">src\main.py</a></td>
+                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html#t77"><data value='str__'>Order.__str__</data></a></td>
                 <td>1</td>
                 <td>0</td>
                 <td>0</td>
@@ -235,10 +243,10 @@
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html"><data value=''><span class='no-noun'>(no function)</span></data></a></td>
-                <td>55</td>
+                <td>50</td>
                 <td>29</td>
                 <td>0</td>
-                <td class="right" data-ratio="26 55">47%</td>
+                <td class="right" data-ratio="21 50">42%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html#t7">src\print_mixin.py</a></td>
@@ -349,10 +357,10 @@
             <tr class="total">
                 <td class="name left">Total</td>
                 <td class="name left">&nbsp;</td>
-                <td>195</td>
+                <td>196</td>
                 <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="154 195">79%</td>
+                <td class="right" data-ratio="155 196">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -364,7 +372,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
     <aside class="hidden">
