+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
 </header>
@@ -99,11 +99,18 @@
                 <td class="right" data-ratio="30 30">100%</td>
             </tr>
             <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_category_iterator_py.html">src\category_iterator.py</a></td>
+                <td>16</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="16 16">100%</td>
+            </tr>
+            <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
-                <td>78</td>
+                <td>63</td>
                 <td>29</td>
                 <td>0</td>
-                <td class="right" data-ratio="49 78">63%</td>
+                <td class="right" data-ratio="34 63">54%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_print_mixin_py.html">src\print_mixin.py</a></td>
@@ -130,10 +137,10 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>195</td>
+                <td>196</td>
                 <td>41</td>
                 <td>0</td>
-                <td class="right" data-ratio="154 195">79%</td>
+                <td class="right" data-ratio="155 196">79%</td>
             </tr>
         </tfoot>
     </table>
@@ -145,7 +152,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.9.2">coverage.py v7.9.2</a>,
-            created at 2025-07-09 12:41 +0300
+            created at 2025-07-09 12:45 +0300
         </p>
     </div>
     <aside class="hidden">
