+++ htmlcov/status.json
@@ -1,1 +1,1 @@
-{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.9.2","globals":"1dd086cc37487ee6006ad316e3600886","files":{"z_145eef247bfb46b6___init___py":{"hash":"e6baa73cda2916dad605215f937a92e1","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_main_py":{"hash":"595a2782d04df28ece70a8db12428f05","index":{"url":"z_145eef247bfb46b6_main_py.html","file":"src\\main.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":78,"n_excluded":0,"n_missing":29,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"1d135d5a5926f61098f7a81695ac4272","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":18,"n_excluded":0,"n_missing":8,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_base_category_order_py":{"hash":"bb8ba3bc179654aa95175bee7c2150e0","index":{"url":"z_145eef247bfb46b6_base_category_order_py.html","file":"src\\base_category_order.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":5,"n_excluded":0,"n_missing":1,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_base_product_py":{"hash":"1a82449f1445021dfec925e35e1697d6","index":{"url":"z_145eef247bfb46b6_base_product_py.html","file":"src\\base_product.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":12,"n_excluded":0,"n_missing":3,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_print_mixin_py":{"hash":"43aeae7892750d78894ba6c293e27236","index":{"url":"z_145eef247bfb46b6_print_mixin_py.html","file":"src\\print_mixin.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":5,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_product_py":{"hash":"6fa575e6c825afd9c9bfd094b5c1d5eb","index":{"url":"z_145eef247bfb46b6_product_py.html","file":"src\\product.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":47,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_category_py":{"hash":"69ab6a59c0b9fb572c8a67f0f0228d96","index":{"url":"z_145eef247bfb46b6_category_py.html","file":"src\\category.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":30,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
+{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.9.2","globals":"1dd086cc37487ee6006ad316e3600886","files":{"z_145eef247bfb46b6___init___py":{"hash":"e6baa73cda2916dad605215f937a92e1","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_main_py":{"hash":"13b6c78b206046a8ad1a79f2e17ddb0d","index":{"url":"z_145eef247bfb46b6_main_py.html","file":"src\\main.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":63,"n_excluded":0,"n_missing":29,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"1d135d5a5926f61098f7a81695ac4272","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":18,"n_excluded":0,"n_missing":8,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_base_category_order_py":{"hash":"bb8ba3bc179654aa95175bee7c2150e0","index":{"url":"z_145eef247bfb46b6_base_category_order_py.html","file":"src\\base_category_order.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":5,"n_excluded":0,"n_missing":1,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_base_product_py":{"hash":"1a82449f1445021dfec925e35e1697d6","index":{"url":"z_145eef247bfb46b6_base_product_py.html","file":"src\\base_product.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":12,"n_excluded":0,"n_missing":3,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_print_mixin_py":{"hash":"43aeae7892750d78894ba6c293e27236","index":{"url":"z_145eef247bfb46b6_print_mixin_py.html","file":"src\\print_mixin.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":5,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_product_py":{"hash":"6fa575e6c825afd9c9bfd094b5c1d5eb","index":{"url":"z_145eef247bfb46b6_product_py.html","file":"src\\product.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":47,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_category_py":{"hash":"69ab6a59c0b9fb572c8a67f0f0228d96","index":{"url":"z_145eef247bfb46b6_category_py.html","file":"src\\category.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":30,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_category_iterator_py":{"hash":"4614b46b430af6e5ccfd92a64003c04e","index":{"url":"z_145eef247bfb46b6_category_iterator_py.html","file":"src\\category_iterator.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":16,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
